requirements.txt
FlightRadar24
pytelegrambotapi
folium
matplotlib
numpy
Pillow
